requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
bitarray==2.9.2
blinker==1.8.2
branca==0.8.0
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.5.7
debugpy==1.8.5
decorator==5.1.1
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
eval_type_backport==0.2.0
executing==2.1.0
filelock==3.16.0
Flask==3.0.3
folium==0.17.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
graphviz==0.20.3
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
ipykernel==6.29.5
ipyplot==1.1.2
ipython==8.27.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.39
langchain-text-splitters==0.2.4
langsmith==0.1.117
lorax-client==0.6.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-enum==1.5.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nodeenv==1.9.1
numpy==1.26.4
openai==1.44.1
orjson==3.10.7
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pdfreader==0.1.15
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.2
predibase==2024.9.2
predibase-api==2024.9.2
progress-table==0.1.26
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.28.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.20.0
pydantic==2.9.1
pydantic_core==2.23.3
pyfiglet==1.0.2
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.8.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-rapidjson==1.20
python-telegram-bot==21.9
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rdflib==7.0.0
referencing==0.35.1
regex==2024.7.24
reportlab==4.2.5
requests==2.32.3
retry==0.9.2
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
selenium==4.27.1
semantic-version==2.10.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.34
stack-data==0.6.3
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
trio==0.28.0
trio-websocket==0.11.1
tritonclient==2.41.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.26.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.0.4
wsproto==1.2.0
xyzservices==2024.9.0
yarl==1.11.1
